Regexp parser and Logic blocks
The Regexp parser can not yet be instanciated to produce any DFA_Trie
(e.g. to extend another parser), the DFA_Trie merge function handle all
cases except the one when n1 branches for a symbol and n2 have an
external klein closure of length 1 with the same symbol (easy fix that
will be done shortly).
The machinery for subrutines/functions for the Logic blocks is currently
under development and omitted.